Refine the flow of getting a seqeunce's parents and children sequences
A sequence's parent names and children names can be extracted from the
location data, and storing parent names and children names to a sequence
object is not necessary. The parents and children attributes of the
Sequence class are replaced with functions, which will return a list
of sequence objects when called.